refactor(#4172): introduce ICompactPipelineTemplateStorage and migrate getPipelineTemplateStorage to typed interface
