+++ streampipes-storage-api/src/main/java/org/apache/streampipes/storage/api/INoSqlStorage.java
@@ -20,7 +20,6 @@ package org.apache.streampipes.storage.api;
 import org.apache.streampipes.model.assets.SpAssetModel;
 import org.apache.streampipes.model.connect.ConnectTransformationScriptTemplate;
 import org.apache.streampipes.model.opcua.Certificate;
-import org.apache.streampipes.model.template.CompactPipelineTemplate;
 
 public interface INoSqlStorage {
 
@@ -76,7 +75,7 @@ public interface INoSqlStorage {
 
   IPrivilegeStorage getPrivilegeStorage();
 
-  CRUDStorage<CompactPipelineTemplate> getPipelineTemplateStorage();
+  ICompactPipelineTemplateStorage getPipelineTemplateStorage();
 
   CRUDStorage<Certificate> getCertificateStorage();
